Job 시스템 구현 후 실행 오류 해결 완료
기존에 레밍 클래스에서 단독으로 구현했던 로직이 Job 시스템으로 정상적으로 똑같이 되는지 테스트 중

+++ src/Digger.h
@@ -5,7 +5,7 @@ class Digger : public Job
 {
 public:
 	void InitAnims(Actor* owner) override;
-	void UpdateStateMachine(int deltaTime) override;
+	void UpdateStateMachine(float deltaTime) override;
 	string GetName() override;
 
 	void SetWalkingRight(bool value);
